+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## 1.1 branch
 ### 1.1.16
 * added `url_to_absolute()` global function;
-* removed `BodyParser::_turnUrlAsAbsolute()` method.
+* removed `BodyParser::_turnUrlAsAbsolute()` and `BodyParser::isHtml()` methods.
 
 ### 1.1.15
 * added `Entity` class.